Added implementation for api-rate-limiter, TDD step 2

+++ utils/__tests__/api-rate-limiter.test.js
@@ -27,20 +27,16 @@ describe('api-rate-limiter tests', () => {
   describe('api-rate-limiter middleware tests', () => {
     const requestWithValidApiKey = httpMocks.createRequest({
       method: 'GET',
-      url: `/weather/uk/london`,
-      headers: {
-        'Authorization': `Basic ${Buffer.from(`${VALID_API_KEY}:x`).toString('base64')}`
-      }
+      url: `/weather/uk/london`
     })
+    requestWithValidApiKey.username = VALID_API_KEY
     const responseForRequestWithValidApiKey = httpMocks.createResponse()
 
     const requestWithInvalidApiKey = httpMocks.createRequest({
       method: 'GET',
-      url: `/weather/uk/london`,
-      headers: {
-        'Authorization': `Basic ${Buffer.from(`${INVALID_API_KEY}:x`).toString('base64')}`
-      }
+      url: `/weather/uk/london`
     })
+    requestWithInvalidApiKey.username = INVALID_API_KEY
     const responseForRequestWithInvalidApiKey = httpMocks.createResponse()
 
     it('should reject invalid API key', async () => {
@@ -65,12 +61,11 @@ describe('api-rate-limiter tests', () => {
 
       expect(spiedNext.called, 'next should be called').to.be.true
       expect(spiedNext.calledWithExactly(), 'next shoud be called with no arguments').to.be.true
-      expect(rateLimitLimit, 'limit should be 1').to.equal(1)
-      expect(rateLimitRemaining, 'remaining should be 0').to.equal(0)
-      expect(rateLimitReset, 'reset should be in the future').to.be.above(nowInUtcSeconds)
+      expect(parseInt(rateLimitLimit, 10), 'limit should be 1').to.equal(1)
+      expect(parseInt(rateLimitRemaining, 10), 'remaining should be 0').to.equal(0)
+      expect(parseInt(rateLimitReset, 10), 'reset should be in the future').to.be.above(nowInUtcSeconds)
     })
-    it('should reject 2nd API call with valid API key', async () => {
-      // Because the limit for test is set to 1 per 60 seconds
+    it('should reject 2nd API call with valid API key, because the limit for test is set to 1 per 60 seconds', async () => {
       const spiedNext = sinon.spy()
       const nowInUtcSeconds = Math.floor(Date.now() / 1000)
       await apiKeyRateLimiter(requestWithValidApiKey, responseForRequestWithValidApiKey, spiedNext)
@@ -85,15 +80,17 @@ describe('api-rate-limiter tests', () => {
 
       expect(spiedNext.called, 'next should be called').to.be.true
       expect(nextCalledWith.name === 'TooManyRequestsError', 'should call next with TooManyRequestsError error').to.be.true
-      expect(rateLimitLimit, 'limit should be 1').to.equal(1)
-      expect(rateLimitRemaining, 'remaining should be 0').to.equal(0)
-      expect(rateLimitReset, 'reset should be in the future').to.be.above(nowInUtcSeconds)
+      expect(parseInt(rateLimitLimit, 10), 'limit should be 1').to.equal(1)
+      expect(parseInt(rateLimitRemaining, 10), 'remaining should be 0').to.equal(0)
+      expect(parseInt(rateLimitReset, 10), 'reset should be in the future').to.be.above(nowInUtcSeconds)
     })
   })
 
   after(async () => {
     // Drop database after test
-    await mongoose.connection.db.dropDatabase()
-    await mongoose.disconnect()
+    if (mongoose.connection.readyState) {
+      await mongoose.connection.db.dropDatabase()
+      await mongoose.disconnect()
+    }
   })
 })
